move all metadata feature building into shared functions
